+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## UnRelease
 
-## [1.2.33-beta.3]
+## [1.2.33]
 
 - pause before source changeing.
 - single apply plugin.