+++ packages/plugins/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## UnRelease
 
+## 1.0.11
+
+- support danmaku
+
+## 1.0.11-alpha.1
+
+- add error fallback for poster
+
 ## 1.0.10-beta.3
 
 - playlist